iis/pyproject.toml
[project]
name = "iis"
version = "2.0.0"
description = "Integrated Intelligence Service (IIS) - Unified Orchestration & Mediation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "mcp>=1.0.0",
    "fastapi-mcp>=0.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.0.0",
    "requests>=2.32.0",
    "PyJWT>=2.10.0",
    "python-multipart>=0.0.21",
    "websockets>=14.0",
    "autogen-agentchat>=0.4.0",
    "autogen-ext[openai]>=0.4.0",
    "RapidFuzz>=3.14.0",
    "pyyaml>=6.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
    "traceloop-sdk>=0.33.0",
    "nemoguardrails>=0.11.0",
    "sqlmodel>=0.0.22",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "core*", "integrations*", "static*"]

[tool.uv]
managed = true

[tool.pyright]
include = ["."]
exclude = ["**/__pycache__", ".venv"]
